// mc/src/store/modules/listen/mutations-types.ts
export enum MutationsTypes {
    AUDIO_REF = "AUDIO_REF",
    AUDIO_STATE = "AUDIO_STATE",
    AUDIO_SONG = "AUDIO_SONG",
    AUDIO_LIST = "AUDIO_LIST",
    AUDIO_INDEX = "AUDIO_INDEX",
    AUDIO_ORDER = "AUDIO_ORDER",
    AUDIO_VOLUME = "AUDIO_VOLUME",
    AUDIO_FOLD = "AUDIO_FOLD",
    AUDIO_CURRENTTIME = "AUDIO_CURRENTTIME",
    AUDIO_DURATION = "AUDIO_DURATION",
    AUDIO_LYRIC = "AUDIO_LYRIC",
    AUDIO_LYRIC_K = "AUDIO_LYRIC_K",
    AUDIO_LYRIC_T = "AUDIO_LYRIC_T",
    AUDIO_NETWORK = "AUDIO_NETWORK",
}